Refactor code
Refactor code to improve performance
