+++ mahautils/multics/sim_results_exporter/io_utils.py
@@ -6,7 +6,6 @@ from typing import Dict, List, Union
 import numpy as np
 
 from mahautils.multics import SimResults
-from mahautils.multics.exceptions import SimResultsDataNotFoundError
 
 
 def decode_base64(base64_str: str) -> str:
@@ -24,32 +23,15 @@ def load_simresults(dash_base64_contents: str, sim_results: SimResults) -> None:
     sim_results.parse()
 
 
-def is_exportable(key: str, sim_results: SimResults):
-    """Returns whether a simulation results variable has data associated with it
-    in the simulation results file"""
-    if sim_results.num_time_steps == 0:
-        return True
-
-    try:
-        data = sim_results.get_data(key, sim_results.get_units(key))
-    except SimResultsDataNotFoundError:
-        return False
-
-    return len(data) == sim_results.num_time_steps
-
-
 def sim_results_to_csv(sim_results: SimResults,
                        config: Dict[str, Dict[str, Union[bool, str]]]) -> str:
     """Saves simulation results data to a CSV file"""
     output_keys: List[str] = []
     output_units: List[str] = []
     output_data: List[np.ndarray] = []
-    for key in sim_results.variables:
+    for key in config:
         export_settings = config.get(key, {'export': False})
         if not export_settings['export']:
-            continue
-
-        if not is_exportable(key, sim_results):
             continue
 
         units = str(export_settings['units'])
